Move the core inquisitor code into its own file
