+++ src/sbd.h
@@ -117,6 +117,7 @@ int watchdog_init(void);
 void sysrq_init(void);
 void watchdog_close(bool disarm);
 struct sbd_context *open_device(const char* devname, int loglevel);
+void open_any_device(void);
 void close_device(struct sbd_context *st);
 signed char cmd2char(const char *cmd);
 void * sector_alloc(void);
